[chore #159102302] : Restructure project dir to fix build issues on Travis CI
