+++ .travis.yml
@@ -1,9 +1,3 @@
-# Set up test coverage reporting
-env:
-  global:
-    - GIT_COMMITTED_AT=$(if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then git log -1 --pretty=format:%ct; else git log -1 --skip 1 --pretty=format:%ct; fi)
-    - CC_TEST_REPORTER_ID=[b4a16e91fba06b8068870160b2dd02cbbd793dd238f151a72b2837cefd33409d]
-
 language: python
 python:
     - "3.6"
